+++ src/lib.rs
@@ -9,5 +9,6 @@ pub mod shared;
 pub use node::primitives::BscPrimitives;
 pub use node::primitives::{BscBlock, BscBlockBody, BscBlobTransactionSidecar};
 mod system_contracts;
+pub use system_contracts::is_invoke_system_contract;
 pub use system_contracts::SLASH_CONTRACT;
 pub use system_contracts::{encode_add_node_ids_call, encode_remove_node_ids_call};